Migrate from RTDB to Firestore PART II

+++ app/src/main/java/com/supercilex/robotscouter/data/client/LinkReceiverActivity.kt
@@ -72,7 +72,7 @@ class LinkReceiverActivity : AppCompatActivity() {
         val teamId = teamPairSplit[0]
         val teamNumber = teamPairSplit[1]
 
-        Team(teamNumber, teamId)
+        Team(teamNumber.toLong(), teamId)
     }
 
     private fun getTemplate(link: Uri): String? = link.getQueryParameter(ID_QUERY)
